+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 For compatibility information see `govukFrontendVersion` and `hmrcFrontendVersion` in
 [LibDependencies](project/LibDependencies.scala)
 
+## [3.28.0] - 2022-09-28
+
+### Changed
+
+- Added the ability to pass a custom `BackLink` to `HmrcLayout` instead of just a `backUrl`
+- Added a helper `BackLink.mimicsBrowserBackButtonViaJavaScript` which mimics the browser Back button
+if the referrer is on the same domain
+- Updated `hmrc-frontend` to v5.10.0
+
+### Compatible with
+
+- [hmrc/hmrc-frontend v5.10.0](https://github.com/hmrc/hmrc-frontend/releases/tag/v5.10.0)
+- [alphagov/govuk-frontend v4.3.1](https://github.com/alphagov/govuk-frontend/releases/tag/v4.3.1)
+
 ## [3.27.0] - 2022-09-23
 
 ### Changed